refactor: refactor injection
